bot-backend/requirements.txt
aiosmtplib==2.0.2
annotated-types==0.5.0
boto3==1.18.62
botocore==1.21.65
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
cssselect==1.2.0
cssutils==2.7.1
email-to==0.1.0
greenlet==2.0.2
http-exceptions==0.2.10
idna==3.4
importlib-metadata==6.8.0
inflect==7.0.0
Jinja2==3.1.2
jmespath==0.10.0
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
premailer==3.10.0
pydantic==2.3.0
pydantic-core==2.6.3
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.5.2
six==1.16.0
sqlacodegen==2.3.0.post1
SQLAlchemy==1.4.26
SQLAlchemy-serializer==1.4.1
typing-extensions==4.7.1
urllib3==1.26.16
zipp==3.16.2
